+++ src/ecommerce/components/Modal.tsx
@@ -3,8 +3,8 @@ import Loader from "./Loader";
 function Modal({ text, btnText, loadButton, func, arg, modal, setModal }) {
 
     return (
-        <div className="fixed inset-0 right-0 left-0 bg-gray-500/75 flex justify-center items-center">
-            <div className="flex flex-col w-96 h-40 bg-white rounded-md justify-center items-center p-2 duraion-400 relative">
+        <div className="fixed inset-0 right-0 left-0 bg-gray-500/75 flex justify-center items-center duration-200">
+            <div className="flex flex-col w-96 h-40 bg-white rounded-md justify-center items-center p-2 relative">
                 <button className="font-bold text-gray-600 px-2 absolute top-0 right-0" onClick={() => setModal(!modal)}>X</button>
                 <p className="text-center text-gray-700">{text}</p>
                 <button
