+++ src/ecommerce/icons/LeftArrowIcon.tsx
@@ -1,0 +1,14 @@
+function LeftArrowIcon({ size }) {
+    return (
+        <svg
+            viewBox="0 0 24 24"
+            fill="#242424"
+            height={size}
+            width={size}
+        >
+            <path d="M4.431 12.822l13 9A1 1 0 0019 21V3a1 1 0 00-1.569-.823l-13 9a1.003 1.003 0 000 1.645z" />
+        </svg>
+    );
+}
+
+export default LeftArrowIcon;
