+++ src/ecommerce/pages/Products.tsx
@@ -3,11 +3,10 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useQuery } from 'react-query';
 import { PRODUCTS_QUERY_KEY } from "../constants";
 import { fetchProducts } from "../services/products";
-import camera from "../../assets/camera-img.png";
 
 import FilterBar from "../components/FilterBar";
-import Loader from "../components/Loader";
-import EditIcon from "../icons/EditIcon";
+import CustomButton from "../components/CustomButton";
+import ProductsList from "../components/ProductsList";
 
 function Products() {
 
@@ -31,61 +30,10 @@ function Products() {
                             <p className="text-transparent">-</p>
                             <p className="text-turquoise">productos</p>
                         </p>
-                        {userRole === "admin" && <button
-                            className="w-fit my-4 self-center border-2 border-turquoise font-bold text-turquoise p-4 rounded-md cursor-pointer hover:bg-turquoise hover:text-white transition duration-150 ease-out hover:ease-in"
-                            onClick={() => navigate("/products/create")}
-                        >
-                            Crear producto
-                        </button>}
+                        {userRole === "admin"
+                            && <CustomButton width="fit" text="Crear producto" bgColor="white" textColor="turquoise" borderColor="turquoise" onClick={() => navigate("/products/create")} />}
                     </div>
-                    <ul className="flex flex-wrap justify-center gap-10 mt-10 h-full w-full">
-                        {
-                            data && data.length > 0 && status === 'success' && data.map((product) => (
-                                <li
-                                    onClick={() => navigate(`/products/${product.id}`)}
-                                    className="relative z-0 flex flex-col justify-center items-center w-52 h-fit rounded-xl border-white overflow-hidden shadow shadow-slate-300 cursor-pointer"
-                                    key={product.id}
-                                >
-                                    <div className="w-52 h-52">
-                                        {userRole === "admin" && <button
-                                            className="absolute z-10 text-white pl-[2.5px] bg-turquoise w-6 h-6 rounded-md absolute right-0"
-                                            onClick={(e) => {
-                                                e.stopPropagation();
-                                                navigate(`/products/edit/${product.id}`)
-                                            }}
-                                        >
-                                            <EditIcon size='20' />
-                                        </button>}
-                                        <img
-                                            className="object-cover w-52 h-52"
-                                            src={product.images[0]}
-                                            alt={product.title}
-                                            onError={(e) => { e.target.src = camera }}
-                                        />
-                                    </div>
-                                    <div className="flex flex-col w-full items-center">
-                                        <p className="w-5/6 bg-white text-gray-700 capitalize text-center truncate">
-                                            {product.title}
-                                        </p>
-                                        <p className="w-5/6 bg-white text-gray-700 capitalize text-center truncate">
-                                            {product.category.name}
-                                        </p>
-                                        <p className="w-5/6 bg-white font-medium text-gray-600 text-center text-lg truncate">
-                                            ${product.price}
-                                        </p>
-                                    </div>
-                                    <button
-                                        className="w-fit font-bold bg-turquoise text-white p-2 m-2 self-center rounded-md cursor-pointer hover:bg-white hover:text-turquoise transition duration-150 ease-out hover:ease-in"
-                                    >
-                                        Agregar al carrito
-                                    </button>
-                                </li>
-                            ))
-                        }
-                        {data && data.length == 0 && status === 'success' && <p className="w-full">No se encontraron productos</p>}
-                        {status === 'loading' && <Loader />}
-                        {status === 'error' && <p>Error al cargar los productos</p>}
-                    </ul>
+                    <ProductsList data={data} status={status} />
                 </div>
             </div>
         </div>
